+++ code/src/components/tabs-screen/TabsPage.tsx
@@ -1,12 +1,16 @@
 "use client"
 import { ThemeContext } from "@/utils/theme/context/ThemeContext"
 import { useContext } from "react"
+import { TabsNav } from "./tabs-nav/tabsNav";
+import { MarkdownProvider } from "@/utils/markdown-parser/context/MarkdownContext";
 
 export const TabsPage = () => {
   const theme = useContext(ThemeContext);
   return (
-    <div className="">
-      DOES THIS WORK OR NOT
-    </div>
+    <MarkdownProvider>
+      <div className="">
+        <TabsNav />
+      </div>
+    </MarkdownProvider>
   )
 }
