+++ code/src/components/tabs-screen/tabs-nav/tabsNav.tsx
@@ -1,0 +1,13 @@
+"use client";
+import { useState } from "react";
+
+export const TabsNav: React.FC = () => {
+  return (
+    <div className="">
+      <h1>CURRENT TAB:</h1>
+      <nav>
+        <ul></ul>
+      </nav>
+    </div>
+  )
+}
