--- code/src/hooks/useTheme.ts
@@ -1,7 +1,0 @@
-import { Theme } from "@/utils/theme/context/ThemeContext";
-import { useEffect, useMemo, useState } from "react"
-
-
-
-
-// detects the default browser theme. I could put some silly malicious code in here :3c
